Metadata_Extraction-Completed & CHANGELOG Added
If applied, the completed first stage of this project will be accessible from the 1_Metadata_Extraction dir. The metadata files at the Dropbox link, accessible from the README file, will be a useful reference throughout the entirety of this project.

The CHANGELOG for the scope of this entire project has also been added to the main dir.